pyproject.toml
[tool.poetry]
package-mode = false
name = "craig-summarizer"
version = "0.1.0"
description = "A discord bot that can take audio summarization from tools such as Craig, and generate a summary of the audio transcription."
authors = ["Adam Barton <firepoisonx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llvmlite = "^0.43.0"
numba = "^0.60.0"
openai-whisper = "^20240930"
google-generativeai = "^0.8.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
